+++ sugar/lib/src/time/weekdays.dart
@@ -1,7 +1,9 @@
 /// Utilities to encode and decode the days in a week into an 8-bit bitfield.
-/// A bitfield represents one week with each bit representing a day. The first bit
-/// in the bitfield is always unused, hence starting with the the day with the largest
-/// value (Sunday) on the 2nd left-most bit.
+/// A bitfield represents one week with each bit representing a day. It is big-endian
+/// with the first bit in the bitfield always being unused.
+///
+/// That is to say, it start with the day with the largest value (Sunday) at the 2nd left-most bit.
+/// For example `0100 0100`, will represent Sunday and Wednesday.
 class Weekdays {
 
   /// Encodes the days of a week into a bitfield. The length of [week] should be less
@@ -23,20 +25,27 @@ class Weekdays {
     return encoded;
   }
 
+
   /// Decodes the bitfield into days, i.e. (1 to 7) of a week. [encoded] should be between 0 and 128.
   ///
   /// Behaviour is undefined if [encoded] is invalid.
   static Iterable<int> decode(int encoded) sync* {
-    int i = 1;
-    for (final value in parse(encoded)) {
-      if (value) {
-        yield i;
+    assert(encoded >= 0 && encoded < 128, 'Packed days is "$encoded", should be between 0 and 127');
+    for (var day = 1; day <= 7; day++) {
+      if (encoded.isOdd) {
+        yield day;
       }
-      i++;
+      encoded >>= 1;
     }
   }
 
+
   /// Parses the bitfield into booleans, i.e. `true` if a digit is 1. [encoded] should be between 0 and 128.
+  /// The returned [Iterable] is little-endian with `true` indicating the presence of a day.
+  ///
+  /// That is so say, the first element in the returned iterable will represent whether Monday is present.
+  /// For example, `[true, false, false true, false, false, false]` will indicate that Monday and Thursday
+  /// is present.
   ///
   /// Behaviour is undefined if [encoded] is invalid.
   static Iterable<bool> parse(int encoded) sync* {
@@ -45,6 +54,27 @@ class Weekdays {
       yield encoded.isOdd;
       encoded >>= 1;
     }
+  }
+
+  /// Encodes the booleans into a bitfield. The length of [week] should be 7.
+  /// [week] is expected to be little-endian with `true` indicating the presence of a day.
+  ///
+  /// That is so say, the first element in [week] will represent whether Monday is present.
+  /// For example, `[true, false, false true, false, false, false]` will indicate that Monday
+  /// and Thursday is present.
+  ///
+  /// Behaviour is undefined if the given week is invalid.
+  static int unparse(Iterable<bool> week) {
+    assert(week.length == 7, 'Number of days is "${week.length}", should be 7');
+    var encoded = 0;
+    for (final day in week.toList().reversed) {
+      encoded <<= 1;
+      if (day) {
+        encoded += 1;
+      }
+    }
+
+    return encoded;
   }
 
 }